Refactor | move IndexOutOfBoundsExceptionDemo to indexoutofbounds package
