+++ core-java-modules/core-java-exceptions-3/src/main/java/com/baeldung/indexoutofbounds/IndexOutOfBoundsExceptionDemo.java
@@ -1,4 +1,4 @@
-package com.baeldung.exception;
+package com.baeldung.indexoutofbounds;
 
 import java.util.ArrayList;
 import java.util.Collections;
