+++ core-java-modules/core-java-exceptions-3/src/test/java/com/baeldung/indexoutofbounds/IndexOutOfBoundsExceptionDemoUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.exception;
+package com.baeldung.indexoutofbounds;
 
 import org.junit.jupiter.api.Test;
